requirements.txt
customtkinter
python-dotenv
mysql-connector-python
pandas
google-generativeai